mapping and standardizing merged
